refactor(cli): extract global-only commands into vp_global_only_cli
Move the env, upgrade, and implode clap definitions out of
crates/vp_global_cli/src/cli.rs into a shared crate as GlobalOnlyCommand,
and flatten it back into the global Commands enum. The crate also exposes
is_global_only(name), derived from the enum through clap, so the names can
be recognised elsewhere without a hand-maintained list that can drift from
the global surface.

The global CLI re-exports the moved types, so commands/env/* is unchanged.
No behavior change; global PTY snapshots pass without updates.
