+++ crates/vp_global_cli/Cargo.toml
@@ -36,6 +36,7 @@ indexmap = { workspace = true }
 indicatif = { workspace = true }
 indoc = { workspace = true }
 vp_error = { workspace = true }
+vp_global_only_cli = { workspace = true }
 vp_js_runtime = { workspace = true }
 vp_pm_cli = { workspace = true }
 vt_path = { workspace = true }
