+++ crates/vp_global_only_cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "vp_global_only_cli"
+version = "0.0.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+publish = false
+rust-version.workspace = true
+
+[dependencies]
+clap = { workspace = true, features = ["derive"] }
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
